added a feature for task get to be able to obtain atomic data for more than 1 atom.
For example, when analyzing the functionality of a charge calculation method, we often need to consider all the atomic charges across all the frames. with the current improvement, atom selection using --atom in get task is not limited to only 1 atom, and multiple atoms or all atoms can be easily accessed and their data can be easily exported in a wide-format table.